test(desktop): exercise window lifecycle bootstrap
